requirements.txt
mcp-agent
fastmcp
openai
tavily-python
pymongo
python-dotenv
